feat: update introduction page & refactor intro-file-path
