+++ components/main-page/introduction/Introduction.module.scss
@@ -32,21 +32,18 @@
   font-size: var(--font-xl);
   font-weight: var(--font-w-xs);
 
-  opacity: 0;
-  transform: translateY(10px);
-  transition: font 0.1s ease-in-out, transform 0.5s ease-in-out;
-
   z-index: 10;
 
   @media (max-width: $tablet-breakpoint) {
     font-size: var(--font-s);
     padding-bottom: 50px;
   }
+}
 
-  &.fade-in {
-    opacity: 1;
-    transform: translateY(0);
-  }
+.focus-line {
+  display: flex;
+
+  flex-wrap: wrap;
 }
 
 .carousel-text-wrapper {
