+++ components/main-page/introduction/hover-text-item/HoverTextItem.module.scss
@@ -1,0 +1,34 @@
+.wrapper {
+  position: relative;
+
+  display: flex;
+  align-items: center;
+
+  margin: 0 0.2rem;
+
+  border-radius: 8px;
+
+  transition: background 0.3s, color 0.3s, padding 0.3s;
+
+  cursor: default;
+}
+
+.wrapper.hovered {
+  background: var(--primary-color-r);
+  color: var(--text-color-r);
+  font-weight: var(--font-w-s);
+
+  padding-left: 0.6rem;
+}
+
+.hovered-emoji {
+  overflow: hidden;
+  width: 0;
+
+  transition: width 0.3s cubic-bezier(0.7, 0, 0.2, 1), margin-left 0.3s;
+  vertical-align: middle;
+}
+.emoji-visible {
+  width: 1.2em;
+  margin-left: 0.2rem;
+}
