+++ components/main-page/introduction/text-auto-slider/TextAutoSlider.module.scss
@@ -23,6 +23,7 @@
   gap: 8px;
 
   color: var(--text-color);
+  font-weight: var(--font-w-s);
 
   svg {
     width: 35px;
